Create 'base' include directory and move relevant headers
