+++ include/adapter/precice_adapter.h
@@ -10,8 +10,8 @@
 #include <deal.II/matrix_free/matrix_free.h>
 
 #include <adapter/dealii_interface.h>
+#include <base/q_equidistant.h>
 #include <precice/SolverInterface.hpp>
-#include <q_equidistant.h>
 
 #include <ostream>
